% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SABCplot.R
\name{SABCplot}
\alias{SABCplot}
\title{S-ABCplot}
\usage{
SABCplot(
  effect_measure = NULL,
  BF_type = NULL,
  estimate = NULL,
  stderr = NULL,
  N1 = NULL,
  N2 = NULL,
  zte = NULL,
  ate = NULL,
  mte = NULL,
  conf_levels = NULL,
  low_axes = NULL,
  upp_axes = NULL,
  length_axes = NULL
)
}
\arguments{
\item{effect_measure}{String indicating the label of the lower horizontal axis."difference in means", "risk difference", "log risk ratio", "log odds ratio", "log hazard ratio"}

\item{BF_type}{String indicating the BF hypotheses setting of plot. "simple", "composite"}

\item{estimate}{Numerical vector containing the estimate.}

\item{N1}{Numerical vector containing the sample size(s). Must be equal the number of estimates.}

\item{N2}{Numerical vector containing the sample size(s). Must be equal the number of estimates.}

\item{zte}{Numerical vector containing zero treatment effect. Must be equal null hypothesis of the null hypothesis statistical test.}

\item{ate}{Numerical vector containing assumed treatment effect. Must be equal treatment effect used in the sample size calculation.}

\item{mte}{(optional)Numerical vector containing minimum treatment effect. Do not write if the BF type is simple.}

\item{conf_levels}{Numerical vector indicating the confidence level(s). Bust be between 0 and 1.}

\item{low_axes}{Numerical vector indicating lower limits of the bottom horizontal axis.}

\item{upp_axes}{Numerical vector indicating upper limits of the bottom horizontal axis.}

\item{length_axes}{Numerical vector indicating scaling of the bottom horizontal axis.}
}
\description{
\code{SABCplot} The function SABCplot is the function that plots Figures 3 and 4.
}
\examples{
=========================================================================================
Graphics in the paper Figure 3, 4
=========================================================================================

adjestimate <- 16.7
adjlower <- 4.2
adjupper <- 29.2
N1 <- 95
N2 <- 95
adjdf <- N1+N2-4
adjse <- (adjestimate - 4.2)/qt((1 - (1 - 0.95)/2), adjdf)
estimate <- adjestimate
se <- adjse
--
Figure3
--
SABCplot(effect_measure = "difference in means"
        , BF_type = "composite"
        , estimate = estimate
        , stderr = se
        , N1 = N1
        , N2 = N2
        , zte = 0
        , ate = 58
        , mte = 8
        , conf_levels = c(0.95)
        , low_axes = -10
        , upp_axes = 70
        , length_axes = 10)
ggsave("ABCplot_mte8.jpeg", dpi = 300, width = 11, height = 9)
--
Figure4
--
SABCplot(effect_measure = "difference in means"
        , BF_type = "composite"
        , estimate = estimate
        , stderr = se
        , N1 = N1
        , N2 = N2
        , zte = 0
        , ate = 58
        , mte = 20
        , conf_levels = c(0.95)
        , low_axes = -10
        , upp_axes = 70
        , length_axes = 10)
ggsave("ABCplot_mte20.jpeg", dpi = 300, width = 11, height = 9)
}
